requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1

matplotlib==3.7.1
setuptools==67.7.2

hydra-core
hydra_colorlog

seaborn
normflows~=1.6
torch==2.0.0
pot
pyro-ppl
pytorch-lightning
mlflow==1.27.0
numpy~=1.24.3
scikit-learn~=1.2.2

omegaconf~=2.3.0
tqdm~=4.65.0
git+https://github.com/fadel/pytorch_ema